UMD Pattern
Implement Universal Module Definition (UMD) Pattern for RevExit. Disable dynamic loading of jquery 2.x, this logic was failing as it was only supposed to source jQuery from CDN when it was not in scope. However it was loading a 2.x branch of jQuery regardless.

Performed JS cleanups based on iOS Simulator console notices.

NOTE: UMD is a pre-requisite for ExitPop 2.0, and is already being utilized in the 2.x series.
Signed-off-by: Julien Chinapen <julien@revcontent.com>